Remove charts from JmxServerConfiguration
We simply ignore property charts of the configuration file.
We keep charts in the configuration class.

Purpose: We plan to add a whitelist and/or blacklist chart configuration of the pool of commonChart configuration of the JmxPlugin.
We think it is better to maintain one list of possible charts and exclude or include them from specific configurations.