Add user documentation (#33)
* Add documentation for release 1.0.0

Document configuration files and schemas.
Simplify other javadoc comments.
Remove not needed comments.  

* JmxChartConfiguration. Rename subsection to family

Descition Reason: Use the same configuration/property names like
netdata uses them. There might be better names in the actual context
but we do not want to irretate users maintaining two names for the same
thing.

* Mute output for successful tests.

Do not print systemOut and systemErr to console for successful tests.

* JmxPluginConfiguration

rename property autoMonitorLocalVirtualMachines to autoDetectLocalVirtualMachines.

Reason: This is more descriptive when reading the configuration.

* Remove charts from JmxServerConfiguration

We simply ignore property charts of the configuration file.
We keep charts in the configuration class.

Purpose: We plan to add a whitelist and/or blacklist chart configuration of the pool of commonChart configuration of the JmxPlugin.
We think it is better to maintain one list of possible charts and exclude or include them from specific configurations.

* Rename JmxChart configuration property dimType to dimensionAlgorithm

Reason: Use the same configuration/property names like
    netdata uses them. There might be better names in the actual context
    but we do not want to irretate users maintaining two names for the same
    thing.